Fix expanded cu_seqlens dtype: must be int32 for flash_attn
flash_attn_varlen_func requires cu_seqlens_k to be int32.
Force dtype=torch.int32 when building expanded_cu_seqlens.

+++ prefix-sharing/prefix_sharing/integrations/verl_attention.py
@@ -182,11 +182,11 @@ def _make_verl_attention_patch(original_forward: Any) -> Any:
             tp_rank=tp_rank,
         )
 
-        # Build expanded cu_seqlens for flash_attn
+        # Build expanded cu_seqlens for flash_attn (must be int32)
         expanded_cu_seqlens = torch.tensor(
             [0] + list(ctx.prefix_sharing_plan.expanded_lengths_kv),
             device=cu_seqlens.device,
-            dtype=cu_seqlens.dtype,
+            dtype=torch.int32,
         ).cumsum(0)
         max_seqlen_expanded = max(ctx.prefix_sharing_plan.expanded_lengths_kv)
